move non negative sphum to vcm
